Provide commonjs alternatives automatically
- TS configs for CJS & ESM separated
- Outputs in two different folders
- CJS package.json stub to override package type and force
commonjs.
- Provide 2 sets of type declarations. This makes it easier to
work with module "node16" config in typescript. Typescript uses the
closest package.json to determine the module type of the type
declarations. So, even if we provide the "types"
keyword in package.json, it can only be right for one type of import:
CJS or MJS, whichever the nearest package.json says.
- package.json exports
- Closes #9